+++ wol.hpp
@@ -1,0 +1,98 @@
+#ifndef WOL_HPP
+#define WOL_HPP
+
+#include <vector>
+#include <string>
+#include <string.h>
+#include <unistd.h>
+#include <algorithm>
+#include <arpa/inet.h>
+
+using namespace std;
+
+extern string response;
+extern pthread_mutex_t *mx;
+
+unsigned char* createMagicPacket(vector<int> mac){
+	unsigned char *packet = (unsigned char*)malloc(102 * sizeof(unsigned char));
+	for(int i = 0; i < 6; i++)
+		packet[i] = 0xff;
+	for(int i = 1; i <= 16; i++)
+		for(int j = 0; j < mac.size(); j++)
+			packet[i*6+j] = mac[j];
+	return packet;
+}
+
+void sendMagicPackage(string ip, vector<int> mac){
+	unsigned char *packet = createMagicPacket(mac);
+	int udpSocket, broadcast = 1;
+	struct sockaddr_in server, client;
+	bzero(&server, sizeof(server));
+	bzero(&client, sizeof(client));
+
+	udpSocket = socket(AF_INET, SOCK_DGRAM, 0);
+
+	if(setsockopt(udpSocket, SOL_SOCKET, SO_BROADCAST, &broadcast, sizeof(broadcast)) == -1){
+		response.clear();
+		response += "\nsetsockopt (SO_BROADCAST)";
+		int x = response.length();
+		pthread_mutex_lock(mx);
+		write(1, &x, sizeof(int));
+		write(1, response.c_str(), x * sizeof(char));
+		pthread_mutex_unlock(mx);
+		exit(EXIT_FAILURE);
+	}
+
+	client.sin_family = AF_INET;
+	client.sin_addr.s_addr = INADDR_ANY;
+	client.sin_port = 0;
+	bind(udpSocket, (struct sockaddr*) &client, sizeof(client));
+
+	server.sin_family = AF_INET;
+	server.sin_addr.s_addr = inet_addr(ip.c_str());
+	server.sin_port = htons(9);
+
+	if(sendto(udpSocket, packet, sizeof(unsigned char) * 102, 0, (struct sockaddr*) &server, sizeof(server)) == 102){
+		response.clear();
+		response += '\n';
+		response += ip;
+		response += ":   wake signal sent";
+		int x = response.length();
+		pthread_mutex_lock(mx);
+		write(1, &x, sizeof(int));
+		write(1, response.c_str(), x * sizeof(char));
+		pthread_mutex_unlock(mx);
+	}
+	close(udpSocket);
+	free(packet);
+}
+
+string getIp(string c){
+	return c.substr(0, c.find(":"));
+}
+
+string getMac(string c){
+	int x = c.find(":");
+	if(x == -1)
+		return "";
+	return c.substr(x + 1);
+}
+
+vector<int> getMacValuesFromString(string mac){
+	vector<int> res(6);
+	transform(mac.begin(), mac.end(), mac.begin(), ::toupper);
+	for(int i = 0; i < 6; i++){
+		if(mac[i * 3] >= 'A')
+			res[i] = mac[i * 3] - 'A' + 10;
+		else
+			res[i] = mac[i * 3] - '0';
+		res[i] *= 16;
+		if(mac[i * 3 + 1] >= 'A')
+			res[i] += mac[i * 3 + 1] - 'A' + 10;
+		else
+			res[i] += mac[i * 3 + 1] - '0';
+	}
+	return res;
+}
+
+#endif
